Improve performance of sortObjListByProp by memoizing sorts.
Details:
Store the previous result for a given returned function, as well as a reference to the objList. If the function is called again with the same objList, return the same result rather than re-sorting the array again.

In addition, slice entire objList, rather than chopping off the first element.